// src/providers/dataTable/interfaces.ts
export type IndexColumnDataType =
  | 'string'
  | 'number'
  | 'date'
  | 'date-time'
  | 'time'
  | 'boolean'
  | 'reference-list-item'
  | 'multiValueRefList'
  | 'entity'
  | 'action'
  | 'other';

export type IndexColumnFilterOption =
  | 'contains'
  | 'startsWith'
  | 'endsWith'
  | 'equals'
  | 'lessThan'
  | 'greaterThan'
  | 'between'
  | 'before'
  | 'after';

export type SortDirection = 0 /*asc*/ | 1 /*desc*/;
export type ColumnSorting = 'asc' | 'desc';

export interface ITableColumn {
  dataFormat?: string;
  id?: string;
  accessor: string;
  header: string;
  isVisible: boolean; // is visible in the table (including columns selector, filter etc.)
  show?: boolean; // is visible on client
  dataType?: IndexColumnDataType;
  filterOption?: IndexColumnFilterOption;
  filter?: any;
  isFilterable: boolean;
  isSortable: boolean;
  isEditable?: boolean;
  defaultSorting?: SortDirection;
  columnId?: string;
  propertyName?: string;
  name?: string;
  caption?: string;
  allowShowHide?: boolean;
  //width?: string;
  referenceListName?: string;
  referenceListModule?: string;
  entityReferenceTypeShortAlias?: string;
  allowInherited?: boolean;

  minWidth?: number;
  maxWidth?: number;
}
// ...
export interface ITableFilter {
  readonly columnId: string;
  readonly filterOption: IndexColumnFilterOption;
  readonly filter: any;
}

export interface IQuickFilter {
  readonly id: string;
  readonly name: string;
  readonly selected?: boolean;
}

export interface ICustomFilter {
  readonly id: string;
  readonly name: string;
  readonly columns: ITableColumn[];
  readonly isPrivate: boolean;
  readonly isApplied?: boolean;
}

export type FilterExpressionType = 'jsonLogic' | 'hql';
export type FilterType = 'predefined' | 'user-defined' | 'quick';
export interface IStoredFilter {
  id: string;

  name: string;

  tooltip?: string;
  // Exclusive filters cannot be applied on top of other filters. Only one can be selected

  isExclusive?: boolean;
  // Private filters are managed within the data table control
  isPrivate?: boolean;

  expressionType?: FilterExpressionType | string;

  expression?: string | object;

  filterType?: string;

  // use
  useExpression?: boolean;

  onlyFetchWhenFullyEvaluated?: boolean;

  selected?: boolean;

  defaultSelected?: boolean;

  //#region dynamic expressions
  hasDynamicExpression?: boolean;

  allFieldsEvaluatedSuccessfully?: boolean;

  unevaluatedExpressions?: string[];
  //#endregion
}